♻️ refactor:  ディレクトリ名を変更
Issue #132
