+++ lib/config/router.dart
@@ -8,7 +8,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 
 import '../presentation/components/repo/repo_detail_view_notifier.dart';
-import '../presentation/pages/error/error_page.dart';
+import '../presentation/pages/common/error_page.dart';
 import '../presentation/pages/repo/repo_index_page.dart';
 import '../presentation/pages/repo/repo_search_page.dart';
 import '../presentation/pages/repo/repo_view_page.dart';
